Particle: Moved resize() function into ParticleSystem class
Having a resize function for just the data vector is needed when resizing the parent particle system in the pdata io.
This way it is possible to read pdata and only afterwards the parent particle system. The resizeAll() function could also be used to resize the data vector of a particle system. However, pdata should always only modify its parent state as there is no relationship between pdata objects, only between pdata and parent particle systems.