Particle: Added maximum particle system size option
Particle systems can now set an upper limit for the number of particles they store.
While this should theoretically not be needed (adjustNumber() should manage the particle count) there are situations (e.g. big external forces resulting in big splashes but limited hardware) where one would like to cap the number of allowed particles in advance. An obvious drawback of having such a field is that there will be cases where particles cannot be sampled because of the cap although they would be very important to have. This will be counterbalanced by sampling particles in a random order. All other cases should not make use of the maximum particle field (leave at 0) or simply use a higher particle cap.