Grid: Made name field accessible via Python
Exposing the name field like this (with a setter) is only be a temporary solution. Ideally, the name field should be accessible in a more pythonic way, i.e. someGrid.name = ...

However, in order to achieve this one has to make deeper changes: There could be a Object class that solvers, grids, etc. inherit from (instead of inheriting from PbClass). This Object class would inherit from PbClass, contain the 'name' field and act as a bridge between solvers, grids, etc. and PbClass.